Move fixtures to start module
